Started writing some address lookup routines.
I started doing this jazz by hand. I fogot that domain names would
be delimited by some magic instead of a period. Because of that,
currently my strncpy is totally not giving me a usable host name!

Gotta go back to the RFC and check out to handle the hostname
compression goodness. Also, later, I need to make sure everything
is robust and easy to follow. It's all just glopped in there right
now, so it's not readable... I'll probably clean it up, then jump
back to this stuff.